+++ api/Services/Interfaces/IPaymentService.cs
@@ -4,6 +4,5 @@ public interface IPaymentService
 {
   Task<string> CreateCheckoutSessionAsync(int userId, int planId);
   Task<string> CreatePortalSessionAsync(string customerId);
-  Task HandleWebhookAsync(string payload, string signature);
   Task<Stripe.Checkout.Session?> VerifyCheckoutSessionAsync(string sessionId);
 }
