+++ api/Services/Interfaces/IWebhookOrchestrator.cs
@@ -1,0 +1,6 @@
+namespace RadioWash.Api.Services.Interfaces;
+
+public interface IWebhookOrchestrator
+{
+    Task HandleWebhookAsync(string payload, string signature);
+}
